allow for a special middle button and allow for custom styling of that button
the idea here is that I can have a custom middle plus button, that has different styling. One case is that I want it to have no flex, and just be a specific size, so that this can allow the left and right tabs take the flex space they need. Otherwise they would be sharing the whole space, and it would seem that they are more pushed to the sides that they should.
You can see this in practice if you have three tabs and have `hasMiddleItem` `false`, and `true`. When `false`, all three tabs share the space. When `true`, the left and right tabs share the space, and the middle one just takes the assigned space its given in the `middleItemStyle`.